+++ examples/showoci/showoci_data.py
@@ -117,6 +117,7 @@ class ShowOCIData(object):
             'config_file': self.service.flags.config_file,
             'config_profile': self.service.flags.config_section,
             'use_instance_principals': self.service.flags.use_instance_principals,
+            'use_delegation_token': self.service.flags.use_delegation_token,
             'version': self.service.flags.showoci_version,
             'override_tenant_id': self.service.flags.filter_by_tenancy_id,
             'datetime': start_time,
@@ -2651,6 +2652,11 @@ class ShowOCIData(object):
             oda = self.service.search_multi_items(self.service.C_DATA_AI, self.service.C_DATA_AI_ODA, 'region_name', region_name, 'compartment_id', compartment['id'])
             if oda:
                 data_ai['oda'] = oda
+
+            # bds
+            bds = self.service.search_multi_items(self.service.C_DATA_AI, self.service.C_DATA_AI_BDS, 'region_name', region_name, 'compartment_id', compartment['id'])
+            if bds:
+                data_ai['bds'] = bds
 
             # data science
             ds = self.service.search_multi_items(self.service.C_DATA_AI, self.service.C_DATA_AI_SCIENCE, 'region_name', region_name, 'compartment_id', compartment['id'])