+++ src/oci/data_flow/models/application_summary.py
@@ -138,7 +138,7 @@ class ApplicationSummary(object):
     def compartment_id(self):
         """
         **[Required]** Gets the compartment_id of this ApplicationSummary.
-        The OCID of the compartment that contains this application.
+        The OCID of a compartment.
 
 
         :return: The compartment_id of this ApplicationSummary.
@@ -150,7 +150,7 @@ class ApplicationSummary(object):
     def compartment_id(self, compartment_id):
         """
         Sets the compartment_id of this ApplicationSummary.
-        The OCID of the compartment that contains this application.
+        The OCID of a compartment.
 
 
         :param compartment_id: The compartment_id of this ApplicationSummary.