+++ src/oci/data_flow/models/run_summary.py
@@ -223,7 +223,7 @@ class RunSummary(object):
     def compartment_id(self):
         """
         **[Required]** Gets the compartment_id of this RunSummary.
-        The OCID of the compartment that contains this application.
+        The OCID of a compartment.
 
 
         :return: The compartment_id of this RunSummary.
@@ -235,7 +235,7 @@ class RunSummary(object):
     def compartment_id(self, compartment_id):
         """
         Sets the compartment_id of this RunSummary.
-        The OCID of the compartment that contains this application.
+        The OCID of a compartment.
 
 
         :param compartment_id: The compartment_id of this RunSummary.