+++ src/oci/streaming/models/stream_pool.py
@@ -71,6 +71,22 @@ class StreamPool(object):
             The value to assign to the kafka_settings property of this StreamPool.
         :type kafka_settings: KafkaSettings
 
+        :param custom_encryption_key:
+            The value to assign to the custom_encryption_key property of this StreamPool.
+        :type custom_encryption_key: CustomEncryptionKey
+
+        :param is_private:
+            The value to assign to the is_private property of this StreamPool.
+        :type is_private: bool
+
+        :param endpoint_fqdn:
+            The value to assign to the endpoint_fqdn property of this StreamPool.
+        :type endpoint_fqdn: str
+
+        :param private_endpoint_settings:
+            The value to assign to the private_endpoint_settings property of this StreamPool.
+        :type private_endpoint_settings: PrivateEndpointSettings
+
         :param freeform_tags:
             The value to assign to the freeform_tags property of this StreamPool.
         :type freeform_tags: dict(str, str)
@@ -88,6 +104,10 @@ class StreamPool(object):
             'lifecycle_state_details': 'str',
             'time_created': 'datetime',
             'kafka_settings': 'KafkaSettings',
+            'custom_encryption_key': 'CustomEncryptionKey',
+            'is_private': 'bool',
+            'endpoint_fqdn': 'str',
+            'private_endpoint_settings': 'PrivateEndpointSettings',
             'freeform_tags': 'dict(str, str)',
             'defined_tags': 'dict(str, dict(str, object))'
         }
@@ -100,6 +120,10 @@ class StreamPool(object):
             'lifecycle_state_details': 'lifecycleStateDetails',
             'time_created': 'timeCreated',
             'kafka_settings': 'kafkaSettings',
+            'custom_encryption_key': 'customEncryptionKey',
+            'is_private': 'isPrivate',
+            'endpoint_fqdn': 'endpointFqdn',
+            'private_endpoint_settings': 'privateEndpointSettings',
             'freeform_tags': 'freeformTags',
             'defined_tags': 'definedTags'
         }
@@ -111,6 +135,10 @@ class StreamPool(object):
         self._lifecycle_state_details = None
         self._time_created = None
         self._kafka_settings = None
+        self._custom_encryption_key = None
+        self._is_private = None
+        self._endpoint_fqdn = None
+        self._private_endpoint_settings = None
         self._freeform_tags = None
         self._defined_tags = None
 
@@ -291,6 +319,100 @@ class StreamPool(object):
         :type: KafkaSettings
         """
         self._kafka_settings = kafka_settings
+
+    @property
+    def custom_encryption_key(self):
+        """
+        **[Required]** Gets the custom_encryption_key of this StreamPool.
+
+        :return: The custom_encryption_key of this StreamPool.
+        :rtype: CustomEncryptionKey
+        """
+        return self._custom_encryption_key
+
+    @custom_encryption_key.setter
+    def custom_encryption_key(self, custom_encryption_key):
+        """
+        Sets the custom_encryption_key of this StreamPool.
+
+        :param custom_encryption_key: The custom_encryption_key of this StreamPool.
+        :type: CustomEncryptionKey
+        """
+        self._custom_encryption_key = custom_encryption_key
+
+    @property
+    def is_private(self):
+        """
+        Gets the is_private of this StreamPool.
+        True if the stream pool is private, false otherwise.
+        If the stream pool is private, the streams inside the stream pool can only be accessed from inside the associated subnetId.
+
+
+        :return: The is_private of this StreamPool.
+        :rtype: bool
+        """
+        return self._is_private
+
+    @is_private.setter
+    def is_private(self, is_private):
+        """
+        Sets the is_private of this StreamPool.
+        True if the stream pool is private, false otherwise.
+        If the stream pool is private, the streams inside the stream pool can only be accessed from inside the associated subnetId.
+
+
+        :param is_private: The is_private of this StreamPool.
+        :type: bool
+        """
+        self._is_private = is_private
+
+    @property
+    def endpoint_fqdn(self):
+        """
+        Gets the endpoint_fqdn of this StreamPool.
+        The FQDN used to access the streams inside the stream pool (same FQDN as the messagesEndpoint attribute of a :class:`Stream` object).
+        If the stream pool is private, the FQDN is customized and can only be accessed from inside the associated subnetId, otherwise the FQDN is publicly resolvable.
+        Depending on which protocol you attempt to use, you need to either prepend https or append the Kafka port.
+
+
+        :return: The endpoint_fqdn of this StreamPool.
+        :rtype: str
+        """
+        return self._endpoint_fqdn
+
+    @endpoint_fqdn.setter
+    def endpoint_fqdn(self, endpoint_fqdn):
+        """
+        Sets the endpoint_fqdn of this StreamPool.
+        The FQDN used to access the streams inside the stream pool (same FQDN as the messagesEndpoint attribute of a :class:`Stream` object).
+        If the stream pool is private, the FQDN is customized and can only be accessed from inside the associated subnetId, otherwise the FQDN is publicly resolvable.
+        Depending on which protocol you attempt to use, you need to either prepend https or append the Kafka port.
+
+
+        :param endpoint_fqdn: The endpoint_fqdn of this StreamPool.
+        :type: str
+        """
+        self._endpoint_fqdn = endpoint_fqdn
+
+    @property
+    def private_endpoint_settings(self):
+        """
+        Gets the private_endpoint_settings of this StreamPool.
+
+        :return: The private_endpoint_settings of this StreamPool.
+        :rtype: PrivateEndpointSettings
+        """
+        return self._private_endpoint_settings
+
+    @private_endpoint_settings.setter
+    def private_endpoint_settings(self, private_endpoint_settings):
+        """
+        Sets the private_endpoint_settings of this StreamPool.
+
+        :param private_endpoint_settings: The private_endpoint_settings of this StreamPool.
+        :type: PrivateEndpointSettings
+        """
+        self._private_endpoint_settings = private_endpoint_settings
 
     @property
     def freeform_tags(self):